[dashboard] move StateHead methods to a lib
